Refactor balance handling and server initialization
- Changed InitBalance type from int to int64 across multiple files for consistency in balance management.
- Updated CreateServer function to accept int64 for initial balance.
- Added SetBalance method to Database for setting account balances.
- Modified PrintDB method to return balances as int64.
- Enhanced SafeTimer to support separate execution and view change timeouts, improving timer management in LinearPBFTNode.

+++ internal/config/config.go
@@ -11,7 +11,7 @@ type Config struct {
 	Nodes       map[string]*NodeEntry   `yaml:"nodes"`
 	Clients     map[string]*ClientEntry `yaml:"clients"`
 	DBDir       string                  `yaml:"db_dir"`
-	InitBalance int                     `yaml:"init_balance"`
+	InitBalance int64                   `yaml:"init_balance"`
 }
 
 // NodeEntry represents a node entry in the config
